requirement.txt
websocket-client
streamlit
alpaca-py
yfinance
pandas
numpy